pmemcheck: add register pmem file macro
There are many changes related to the new register macro. The tool will
be more flexible with the separate macro. Some test changes had to be
made to account for the changes. The tool also had a bug in improper
flush detection. All output values in log mode have been changed to
hexadecimal format.